// lib/api/__generated__/models/LoginUser.ts
/* tslint:disable */
/* eslint-disable */
/**
 * RealWorld Conduit API
 * Conduit API documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface LoginUser
 */
export interface LoginUser {
    /**
     * 
     * @type {string}
     * @memberof LoginUser
     */
    email: string;
    /**
     * 
     * @type {string}
     * @memberof LoginUser
     */
    password: string;
}

/**
 * Check if a given object implements the LoginUser interface.
 */
export function instanceOfLoginUser(value: object): boolean {
    if (!('email' in value)) return false;
    if (!('password' in value)) return false;
    return true;
}

export function LoginUserFromJSON(json: any): LoginUser {
    return LoginUserFromJSONTyped(json, false);
}

export function LoginUserFromJSONTyped(json: any, ignoreDiscriminator: boolean): LoginUser {
    if (json == null) {
        return json;
    }
    return {
        
        'email': json['email'],
        'password': json['password'],
    };
}

export function LoginUserToJSON(value?: LoginUser | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'email': value['email'],
        'password': value['password'],
    };
}